refactor: change Game package utils to baseball
