+++ src/main/java/baseball/Game.java
@@ -1,7 +1,8 @@
-package utils;
+package baseball;
 
 import number.Number;
 import number.Numbers;
+import utils.InputUtils;
 
 import java.util.ArrayList;
 import java.util.List;
