+++ src/test/java/number/NumbersTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import utils.Game;
+import baseball.Game;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
