+++ src/test/java/utils/InputUtilsTest.java
@@ -1,5 +1,6 @@
 package utils;
 
+import baseball.Game;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
